requirements.txt
# Stealth Accumulation Scanner AI Agent
# Requirements for Indian Stock Market Analysis

# Data Handling
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
ta>=0.10.0

# HTTP & API
requests>=2.31.0

# Scheduler
apscheduler>=3.10.0

# Telegram Bot
python-telegram-bot>=20.0

# Configuration
pyyaml>=6.0

# Logging
python-dotenv>=1.0.0

# Date/Time handling
python-dateutil>=2.8.0

# LLM/AI Providers (optional - install only the ones you need)
# OpenAI GPT
openai>=1.0.0
# Anthropic Claude
anthropic>=0.25.0
# Google Gemini
google-generativeai>=0.5.0
